+++ src/CosmosDB/CosmosDB.Test/ScenarioTests/MongoOperationsTests.cs
@@ -50,5 +50,19 @@ namespace Microsoft.Azure.Commands.CosmosDB.Test.ScenarioTests.ScenarioTest
         {
             TestRunner.RunTestScript("Test-MongoMigrateThroughputCmdlets");
         }
+
+        [Fact(Skip = "Cannot acquire token credential for a specific audience. No support from test framework. I have verified the tests manually.")]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestMongoDBCollectionThroughputRedistributionCmdlet()
+        {
+            TestRunner.RunTestScript("Test-MongoDBCollectionAdaptiveRUCmdlets");
+        }
+
+        [Fact(Skip = "Cannot acquire token credential for a specific audience. No support from test framework. I have verified the tests manually.")]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestMongoDBCollectionMergeCmdlet()
+        {
+            TestRunner.RunTestScript("Test-MongoDBCollectionMergeCmdlet");
+        }
     }
 }